Rename tests/json to tests/json_gen (fixes #1002)
